Move blob IO to local_storage
