+++ src/ert/storage/local_ensemble.py
@@ -1354,11 +1354,11 @@ class LocalEnsemble(BaseMode):
         blob_type: BlobType | None = None,
     ) -> list[BlobStorageData]:
         """List blob metadata, filtered by type."""
-        return BlobStorageData.load_all(self._path / BLOB_DATA_DIR, blob_type)
+        return self._storage.load_blobs(self._path / BLOB_DATA_DIR, blob_type)
 
     def load_blob(self, uri: str) -> bytes:
         """Load blob bytes by URI."""
-        return BlobStorageData.read_bytes(self._path / BLOB_DATA_DIR, uri)
+        return self._storage.load_blob(self._path / BLOB_DATA_DIR, uri)
 
     @require_write
     def save_blob(
@@ -1370,7 +1370,7 @@ class LocalEnsemble(BaseMode):
             file_type = (
                 "application/x-npz" if blob_event.sparse else "application/x-npy"
             )
-            BlobStorageData.save_blob(
+            self._storage.save_blob(
                 name=blob_event.name,
                 data=blob_event.matrix_bytes,
                 blob_info=MatrixStorageData(
@@ -1381,11 +1381,10 @@ class LocalEnsemble(BaseMode):
                     parameter_group_sizes=blob_event.parameter_group_sizes,
                 ),
                 file_type=file_type,
-                storage=self._storage,
                 blob_dir=blob_dir,
             )
         elif blob_event.event_type == "AnalysisScalingEvent":
-            BlobStorageData.save_blob(
+            self._storage.save_blob(
                 name="scaling_factors",
                 data=blob_event.scaling_bytes,
                 blob_info=ScalingFactorsData(
@@ -1394,7 +1393,6 @@ class LocalEnsemble(BaseMode):
                     num_groups=blob_event.num_groups,
                 ),
                 file_type="application/parquet",
-                storage=self._storage,
                 blob_dir=blob_dir,
             )
         else:
@@ -1404,14 +1402,13 @@ class LocalEnsemble(BaseMode):
                 schema=blob_event.data.header,
                 orient="row",
             ).write_parquet(buf)
-            BlobStorageData.save_blob(
+            self._storage.save_blob(
                 name="observation_report",
                 data=buf.getvalue(),
                 blob_info=ObservationReportData(
                     update_algorithm=blob_event.update_algorithm,
                 ),
                 file_type="application/parquet",
-                storage=self._storage,
                 blob_dir=blob_dir,
             )
 
@@ -1424,12 +1421,11 @@ class LocalEnsemble(BaseMode):
             buf = io.BytesIO()
             df.write_parquet(buf)
             data = buf.getvalue()
-            BlobStorageData.save_blob(
+            self._storage.save_blob(
                 name=df_name,
                 data=data,
                 blob_info=EverestBatchData(dataframe_name=df_name),
                 file_type="application/parquet",
-                storage=self._storage,
                 blob_dir=blob_dir,
             )
